requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
distro==1.9.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
neo4j==5.20.0
openai==0.28.0
PRGenius==1.0.1
psycopg2-binary==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.30
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
